BUG/MINOR: ring: b_peek_varint() returns a uint64_t, not a size_t
The difference matters when building on 32-bit architectures and a
warning was rightfully emitted.

No backport is needed.

+++ src/ring.c
@@ -117,7 +117,7 @@ ssize_t ring_write(struct ring *ring, size_t maxlen, const struct ist pfx[], siz
 	struct appctx *appctx;
 	size_t totlen = 0;
 	size_t lenlen;
-	size_t dellen;
+	uint64_t dellen;
 	int dellenlen;
 	ssize_t sent = 0;
 	int i;
